requirements.txt
requests
lxml
pillow
pyglet
